move common IDE code to a separate file + reorganize things a bit

+++ kernel/fs/fs.h
@@ -32,9 +32,9 @@ struct k_fs_vol_t *k_fs_FormatPartition(struct k_fs_part_t *partition, uint32_t 
 
 void k_fs_FormatDisk(struct k_dsk_disk_t *disk, uint32_t part_table_type);
 
-void k_fs_ReadVolumeBytes(struct k_fs_vol_t *volume, uint32_t block_size, uint32_t first_block, uint32_t offset, uint32_t size, void *buffer); 
+void k_fs_ReadVolumeBytes(struct k_fs_vol_t *volume, uint32_t block_size, uint64_t first_block, uint64_t offset, uint64_t size, void *buffer); 
 
-void k_fs_ReadVolumeBlocks(struct k_fs_vol_t *volume, uint32_t block_size, uint32_t first_block, uint32_t block_count, void *buffer);
+void k_fs_ReadVolumeBlocks(struct k_fs_vol_t *volume, uint32_t block_size, uint64_t first_block, uint64_t block_count, void *buffer);
 
 void k_fs_WriteVolumeBytes(struct k_fs_vol_t *volume, uint32_t block_size, uint32_t first_block, uint32_t offset, uint32_t size, void *buffer);
